+++ src/services/key/KeyIdCollection.php
@@ -1,0 +1,15 @@
+<?php
+namespace PharIo\Phive;
+
+class KeyIdCollection {
+
+    private $keyIds = [];
+
+    public function addKeyId($keyId) {
+        $this->keyIds[] = $keyId;
+    }
+
+    public function has($keyId) {
+        return in_array($keyId, $this->keyIds, true);
+    }
+}